requirements.txt
opencv-python==4.4.0.44
tensorflow-addons==0.10.0
trianglesolver==1.2
dotmap==1.3.23
mtcnn==0.1.1
numpy==1.23.5
matplotlib==3.7.1
tqdm==4.64.1
dlib==19.24.1